% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao.R
\name{calculo}
\alias{calculo}
\title{Calculo}
\usage{
calculo(x, y, op)
}
\arguments{
\item{x}{primeira variavel}

\item{y}{segunda variavel}

\item{op}{operação matemática}
}
\value{
Um valor numérico
}
\description{
Pacote para calculos matemáticos
}
